requirements.txt
pycryptodome>=3.19.0
asyncio>=3.4.3
python-dotenv>=1.0.0 